removed match_islands, renamed file to snapIsland.
